# docs/swagger.yaml
...
info:
  contact: {}
paths:
...
  /workshop:
    get:
      consumes:
      - application/json
      description: Dapatkan semua workshop atau workshop berdasarkan ID tertentu
      parameters:
      - description: ID Workshop
        in: query
        name: id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties: true
            type: object
      summary: Dapatkan semua workshop atau workshop berdasarkan ID tertentu
      tags:
      - Workshop
    post:
      consumes:
      - application/json
      description: Buat workshop baru dengan informasi yang diberikan
      parameters:
      - description: File gambar
        in: formData
        name: file
        required: true
        type: file
      - description: Title
        in: formData
        name: Title
        required: true
        type: string
      - description: Desc
        in: formData
        name: Desc
        required: true
        type: string
      - description: Date
        in: formData
        name: Date
        required: true
        type: string
      - description: Url
        in: formData
        name: Url
        required: true
        type: string
      - description: IsOpen
        in: formData
        name: IsOpen
        required: true
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Buat workshop baru
      tags:
      - Workshop
  /workshop/{id}:
    delete:
      consumes:
      - application/json
      description: Hapus workshop berdasarkan ID yang diberikan
      parameters:
      - description: ID Workshop
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Hapus workshop berdasarkan ID
      tags:
      - Workshop
    get:
      consumes:
      - application/json
      description: Dapatkan satu workshop berdasarkan ID yang diberikan
      parameters:
      - description: ID Workshop
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties: true
            type: object
      summary: Dapatkan satu workshop berdasarkan ID
      tags:
      - Workshop
    put:
      consumes:
      - application/json
      description: Update workshop berdasarkan ID yang diberikan
      parameters:
      - description: ID Workshop
        in: path
        name: id
        required: true
        type: integer
      - description: File gambar
        in: formData
        name: file
        required: true
        type: file
      - description: title
        in: formData
        name: Title
        required: true
        type: string
      - description: desc
        in: formData
        name: Desc
        required: true
        type: string
      - description: date
        in: formData
        name: Date
        required: true
        type: string
      - description: url
        in: formData
        name: Url
        required: true
        type: string
      - description: IsOpen
        in: formData
        name: IsOpen
        required: true
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update workshop berdasarkan ID
      tags:
      - Workshop
swagger: "2.0"