# handler/workshop.go
package handler

import (
	"bytes"
	"context"
	"fmt"
	endpointcount "greenwelfare/endpointCount"
	"greenwelfare/helper"
	"greenwelfare/imagekits"
	"greenwelfare/workshop"
	"io"
	"net/http"
	"strconv"

	"github.com/gin-gonic/gin"
)

type workshopHandler struct {
	workshopService workshop.Service
	endpointService endpointcount.StatisticsService
}

func NewWorkshopHandler(workshopService workshop.Service, endpointService endpointcount.StatisticsService) *workshopHandler {
	return &workshopHandler{workshopService, endpointService}
}

// @Summary Buat workshop baru
// @Description Buat workshop baru dengan informasi yang diberikan
// @Accept json
// @Produce json
// @Security BearerAuth
// @Tags Workshop
// @Param file formData file true "File gambar"
// @Param Title formData string true "Title"
// @Param Desc formData string true "Desc"
// @Param Date formData string true "Date"
// @Param Url formData string true "Url"
// @Param IsOpen formData boolean true "IsOpen"
// @Success 200 {object} map[string]interface{}
// @Failure 400 {object} map[string]interface{}
// @Failure 422 {object} map[string]interface{}
// @Router /workshop [post]
func (h *workshopHandler) CreateWorkshop(c *gin.Context) {
	file, _ := c.FormFile("file")
	src, err := file.Open()
	if err != nil {
		fmt.Printf("error when open file %v", err)
	}
	defer src.Close()

	buf := bytes.NewBuffer(nil)
	if _, err := io.Copy(buf, src); err != nil {
		fmt.Printf("error read file %v", err)
		return
	}

	img, err := imagekits.Base64toEncode(buf.Bytes())
	if err != nil {
		fmt.Println("error reading image ", err)
	}

	fmt.Println("image base 64 format : ", img)

	imageKitURL, err := imagekits.ImageKit(context.Background(), img)
	if err != nil {
		// Tangani jika terjadi kesalahan saat upload gambar
		// Misalnya, Anda dapat mengembalikan respon error ke klien jika diperlukan
		response := helper.APIresponse(http.StatusInternalServerError, "Failed to upload image")
		c.JSON(http.StatusInternalServerError, response)
		return
	}

	var input workshop.CreateWorkshop

	err = c.ShouldBind(&input)

	if err != nil {
		errors := helper.FormatValidationError(err)
		errorMessage := gin.H{"errors": errors}
		response := helper.APIresponse(http.StatusUnprocessableEntity, errorMessage)
		c.JSON(http.StatusUnprocessableEntity, response)
		return
	}

	if err != nil {
		
		response := helper.APIresponse(http.StatusUnprocessableEntity, err.Error())
		c.JSON(http.StatusUnprocessableEntity, response)
		return
	}

	_, err = h.workshopService.CreateWorkshop(input, imageKitURL)
	if err != nil {
		// data := gin.H{"is_uploaded": false}
		response := helper.APIresponse(http.StatusUnprocessableEntity, err.Error())
		c.JSON(http.StatusUnprocessableEntity, response)
		return
	}
	data := gin.H{"is_uploaded": true}
	response := helper.APIresponse(http.StatusOK, data)
	c.JSON(http.StatusOK, response)
}

// @Summary Dapatkan satu workshop berdasarkan ID
// @Description Dapatkan satu workshop berdasarkan ID yang diberikan
// @Accept json
// @Produce json
// @Tags Workshop
// @Param id path int true "ID Workshop"
// @Success 200 {object} map[string]interface{}
// @Failure 400 {object} map[string]interface{}
// @Failure 422 {object} map[string]interface{}
// @Router /workshop/{id} [get]
func (h *workshopHandler) GetOneWorkshop(c *gin.Context) {
	var input workshop.GetWorkshop

	err := c.ShouldBindUri(&input)
	if err != nil {
		errors := helper.FormatValidationError(err)
		errorMessage := gin.H{"errors": errors}
		response := helper.APIresponse(http.StatusUnprocessableEntity, errorMessage)
		c.JSON(http.StatusUnprocessableEntity, response)
		return
	}

	newDel, err := h.workshopService.GetOneWorkshop(input.ID)
	if err != nil {
		
		response := helper.APIresponse(http.StatusUnprocessableEntity, err.Error())
		c.JSON(http.StatusUnprocessableEntity, response)
		return

	}

	userAgent := c.GetHeader("User-Agent")

	err = h.endpointService.IncrementCount("GetByIDWorkshop /Workshop/GetByIDWorkshop", userAgent)
    if err != nil {
        response := helper.APIresponse(http.StatusUnprocessableEntity, err.Error())
		c.JSON(http.StatusUnprocessableEntity, response)
		return
    }

	response := helper.APIresponse(http.StatusOK, newDel)
	c.JSON(http.StatusOK, response)

}

// @Summary Dapatkan semua workshop atau workshop berdasarkan ID tertentu
// @Description Dapatkan semua workshop atau workshop berdasarkan ID tertentu
// @Accept json
// @Produce json
// @Tags Workshop
// @Param id query int false "ID Workshop"
// @Success 200 {object} map[string]interface{}
// @Failure 400 {object} map[string]interface{}
// @Failure 422 {object} map[string]interface{}
// @Router /workshop [get]
func (h *workshopHandler) GetAllWorkshop(c *gin.Context) {
	input, _ := strconv.Atoi(c.Query("id"))

	newBerita, err := h.workshopService.GetAllWorkshop(input)
	if err != nil {
		response := helper.APIresponse(http.StatusUnprocessableEntity, err.Error())
		c.JSON(http.StatusUnprocessableEntity, response)
		return
	}
	userAgent := c.GetHeader("User-Agent")

	err = h.endpointService.IncrementCount("GetAllWorkshop /Workshop/GetAllWorkshop", userAgent)
    if err != nil {
        response := helper.APIresponse(http.StatusUnprocessableEntity, err.Error())
		c.JSON(http.StatusUnprocessableEntity, response)
		return
    }

	response := helper.APIresponse(http.StatusOK, newBerita)
	c.JSON(http.StatusOK, response)
}

// @Summary Update workshop berdasarkan ID
// @Description Update workshop berdasarkan ID yang diberikan
// @Accept json
// @Produce json
// @Security BearerAuth
// @Tags Workshop
// @Param id path int true "ID Workshop"
// @Param file formData file true "File gambar"
// @Param Title formData string true "title"
// @Param Desc formData string true "desc"
// @Param Date formData string true "date"
// @Param Url formData string true "url"
// @Param IsOpen formData boolean true "IsOpen"
// @Success 200 {object} map[string]interface{}
// @Failure 400 {object} map[string]interface{}
// @Failure 422 {object} map[string]interface{}
// @Router /workshop/{id} [put]
func (h *workshopHandler) UpdateWorkshop(c *gin.Context) {
	file, _ := c.FormFile("file")
	src, err := file.Open()
	if err != nil {
		fmt.Printf("error when open file %v", err)
	}
	defer src.Close()

	buf := bytes.NewBuffer(nil)
	if _, err := io.Copy(buf, src); err != nil {
		fmt.Printf("error read file %v", err)
		return
	}

	img, err := imagekits.Base64toEncode(buf.Bytes())
	if err != nil {
		fmt.Println("error reading image ", err)
	}

	fmt.Println("image base 64 format : ", img)

	imageKitURL, err := imagekits.ImageKit(context.Background(), img)
	if err != nil {
		// Tangani jika terjadi kesalahan saat upload gambar
		// Misalnya, Anda dapat mengembalikan respon error ke klien jika diperlukan
		response := helper.APIresponse(http.StatusInternalServerError, "Failed to upload image")
		c.JSON(http.StatusInternalServerError, response)
		return
	}

	var inputID workshop.GetWorkshop
	err = c.ShouldBindUri(&inputID)
	if err != nil {
		errors := helper.FormatValidationError(err)
		errorMessage := gin.H{"errors": errors}
		response := helper.APIresponse(http.StatusUnprocessableEntity, errorMessage)
		c.JSON(http.StatusUnprocessableEntity, response)
		return
	}

	var input workshop.CreateWorkshop
	err = c.ShouldBind(&input)
	if err != nil {
		errors := helper.FormatValidationError(err)
		errorMessage := gin.H{"errors": errors}
		response := helper.APIresponse(http.StatusUnprocessableEntity, errorMessage)
		c.JSON(http.StatusUnprocessableEntity, response)
		return
	}

	workshop, err := h.workshopService.UpdateWorkshop(inputID, input, imageKitURL)
	if err != nil {
		
		response := helper.APIresponse(http.StatusUnprocessableEntity, err.Error())
		c.JSON(http.StatusUnprocessableEntity, response)
		return
	}

	response := helper.APIresponse(http.StatusOK, workshop)
	c.JSON(http.StatusOK, response)
}

// @Summary Hapus workshop berdasarkan ID
// @Description Hapus workshop berdasarkan ID yang diberikan
// @Accept json
// @Produce json
// @Security BearerAuth
// @Tags Workshop
// @Param id path int true "ID Workshop"
// @Success 200 {object} map[string]interface{}
// @Failure 400 {object} map[string]interface{}
// @Failure 422 {object} map[string]interface{}
// @Router /workshop/{id} [delete]
func (h *workshopHandler) DeleteWorkshop(c *gin.Context) {
	var input workshop.GetWorkshop

	err := c.ShouldBindUri(&input)
	if err != nil {
		errors := helper.FormatValidationError(err)
		errorMessage := gin.H{"errors": errors}
		response := helper.APIresponse(http.StatusUnprocessableEntity, errorMessage)
		c.JSON(http.StatusUnprocessableEntity, response)
		return
	}

	newDel, err := h.workshopService.DeleteWorkshop(input.ID)
	if err != nil {
		
		response := helper.APIresponse(http.StatusUnprocessableEntity, err.Error())
		c.JSON(http.StatusUnprocessableEntity, response)
		return

	}
	response := helper.APIresponse(http.StatusOK, newDel)
	c.JSON(http.StatusOK, response)
}